Refactor images (#67)
