+++ frontend/src/img/retouching-examples/after-1.jpg
